orchestrator_service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
tenacity
structlog
asyncpg
redis
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
python-dotenv
prometheus-client
requests
chromadb
beautifulsoup4
